+++ config/settings/prod.py
@@ -63,7 +63,7 @@ CELERY_TASK_QUEUES = (
 # logging
 LOGGING = {
     'version': 1,
-    'disable_existing_loggers': True,
+    'disable_existing_loggers': False,  # use the default
     'root': {
         'level': 'WARNING',
         'handlers': ['sentry']