+++ seed/static/seed/js/controllers/inventory_list_controller.js
@@ -11,6 +11,7 @@ angular.module('BE.seed.controller.inventory_list', [])
     'inventory_service',
     'label_service',
     'data_quality_service',
+    'user_service',
     'inventory',
     'cycles',
     'labels',
@@ -25,6 +26,7 @@ angular.module('BE.seed.controller.inventory_list', [])
               inventory_service,
               label_service,
               data_quality_service,
+              user_service,
               inventory,
               cycles,
               labels,
@@ -175,7 +177,8 @@ angular.module('BE.seed.controller.inventory_list', [])
                 },
                 name: _.constant(null),
                 uploaded: _.constant(null),
-                importFileId: _.constant(response.progress_key.split(':').pop())
+                importFileId: _.constant(response.progress_key.split(':').pop()),
+                orgId: _.constant(user_service.get_organization().id)
               }
             });
             modalInstance.result.then(function () {
@@ -436,9 +439,25 @@ angular.module('BE.seed.controller.inventory_list', [])
           templateUrl: urls.static_url + 'seed/partials/export_inventory_modal.html',
           controller: 'export_inventory_modal_controller',
           resolve: {
-            gridApi: function () {
-              return $scope.gridApi;
+            cycle_id: function () {
+              return $scope.cycle.selected_cycle.id;
+            },
+            ids: function () {
+              var visibleRowIds = _.map($scope.gridApi.core.getVisibleRows($scope.gridApi.grid), function (row) {
+                return row.entity.id;
+              });
+              var selectedRowIds = _.map($scope.gridApi.selection.getSelectedRows(), 'id');
+              return _.filter(visibleRowIds, function (id) {
+                return _.includes(selectedRowIds, id);
+              });
+            },
+            columns: function () {
+              return _.map($scope.columns, 'name');
+            },
+            inventory_type: function () {
+              return $scope.inventory_type;
             }
+
           }
         });
       };