+++ seed/views/properties.py
@@ -8,23 +8,18 @@ All rights reserved.  # NOQA
 :author
 """
 
-# Imports from Standard Library
 import collections
 import re
 from os import path
 
-# Imports from Django
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
-from django.forms.models import model_to_dict
 from django.http import JsonResponse
-from django.utils.timezone import make_naive
 from rest_framework import status
 from rest_framework.decorators import detail_route, list_route
 from rest_framework.renderers import JSONRenderer
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 
-# Local Imports
 from seed.decorators import ajax_request_class
 from seed.filtersets import PropertyViewFilterSet, PropertyStateFilterSet
 from seed.lib.superperms.orgs.decorators import has_perm_class
@@ -67,7 +62,6 @@ from seed.utils.viewsets import (
     SEEDOrgModelViewSet
 )
 
-# Constants
 # Global toggle that controls whether or not to display the raw extra
 # data fields in the columns returned for the view.
 DISPLAY_RAW_EXTRADATA = True
@@ -252,123 +246,8 @@ class PropertyViewSet(GenericViewSet):
                 'total': paginator.count
             },
             'cycle_id': cycle.id,
-            'results': []
+            'results': TaxLotProperty.get_related(property_views, columns)
         }
-
-        # Ids of propertyviews to look up in m2m
-        prop_ids = [property_dict.pk for property_dict in property_views]
-        joins = TaxLotProperty.objects.filter(property_view_id__in=prop_ids)
-
-        # Get all ids of tax lots on these joins
-        taxlot_view_ids = [j.taxlot_view_id for j in joins]
-
-        # Get all tax lot views that are related
-        taxlot_views = TaxLotView.objects.select_related('taxlot', 'state', 'cycle').filter(
-            pk__in=taxlot_view_ids)
-
-        # Map tax lot view id to tax lot view's state data, so we can reference these easily and save some queries.
-        db_columns = Column.retrieve_db_fields()
-        for lot in taxlot_views:
-            # Each object in the response is built from the state data, with related data added on.
-            model_to_dict(lot.state, exclude=['extra_data'])
-
-            for extra_data_field, extra_data_value in lot.state.extra_data.items():
-                if extra_data_field == 'id':
-                    extra_data_field += '_extra'
-
-                # Check if the extra data field is already a database field
-                while extra_data_field in db_columns:
-                    extra_data_field += '_extra'
-
-        taxlot_map = {}
-        for taxlot_view in taxlot_views:
-            taxlot_dict = model_to_dict(taxlot_view.state, exclude=['extra_data'])
-            taxlot_dict['taxlot_state_id'] = taxlot_view.state.id
-
-            # Add extra data fields right to this object.
-            for extra_data_field, extra_data_value in taxlot_view.state.extra_data.items():
-                if extra_data_field == 'id':
-                    extra_data_field += '_extra'
-
-                while extra_data_field in db_columns:
-                    extra_data_field += '_extra'
-
-                taxlot_dict[extra_data_field] = extra_data_value
-
-            # Only return the requested rows. speeds up the json string time.
-            # The front end requests for related columns have 'tax_' prepended to them, so check
-            # for that too.
-            taxlot_dict = {key: value for key, value in taxlot_dict.items() if (key in columns) or
-                           ("tax_{}".format(key) in columns)}
-            taxlot_map[taxlot_view.pk] = taxlot_dict
-            # Replace taxlot_view id with taxlot id
-            taxlot_map[taxlot_view.pk]['id'] = taxlot_view.taxlot.id
-
-        # A mapping of property view pk to a list of taxlot state info for a taxlot view
-        join_map = {}
-        for join in joins:
-            join_dict = taxlot_map[join.taxlot_view_id].copy()
-            join_dict.update({
-                'primary': 'P' if join.primary else 'S',
-                'taxlot_view_id': join.taxlot_view_id
-            })
-            try:
-                join_map[join.property_view_id].append(join_dict)
-            except KeyError:
-                join_map[join.property_view_id] = [join_dict]
-
-        for prop in property_views:
-            # Each object in the response is built from the state data, with related data added on.
-            property_dict = model_to_dict(prop.state, exclude=['extra_data'])
-
-            for extra_data_field, extra_data_value in prop.state.extra_data.items():
-                if extra_data_field == 'id':
-                    extra_data_field += '_extra'
-
-                while extra_data_field in db_columns:
-                    extra_data_field += '_extra'
-
-                property_dict[extra_data_field] = extra_data_value
-
-            # Use property_id instead of default (state_id)
-            property_dict['id'] = prop.property_id
-
-            property_dict['property_state_id'] = prop.state.id
-            property_dict['property_view_id'] = prop.id
-
-            property_dict['campus'] = prop.property.campus
-
-            # All the related tax lot states.
-            property_dict['related'] = join_map.get(prop.pk, [])
-
-            # fix specific time stamps - total hack right now. Need to reconcile with
-            # /data_importer/views.py
-            if property_dict.get('recent_sale_date'):
-                property_dict['recent_sale_date'] = make_naive(
-                    property_dict['recent_sale_date']).isoformat()
-
-            if property_dict.get('release_date'):
-                property_dict['release_date'] = make_naive(
-                    property_dict['release_date']).isoformat()
-
-            if property_dict.get('generation_date'):
-                property_dict['generation_date'] = make_naive(
-                    property_dict['generation_date']).isoformat()
-
-            if property_dict.get('analysis_start_time'):
-                property_dict['analysis_start_time'] = make_naive(
-                    property_dict['analysis_start_time']).isoformat()
-
-            if property_dict.get('analysis_end_time'):
-                property_dict['analysis_end_time'] = make_naive(
-                    property_dict['analysis_end_time']).isoformat()
-
-            # remove the measures from this view for now. Really need to get serializers working
-            # correctly for this class.
-            if property_dict.get('measures'):
-                del property_dict['measures']
-
-            response['results'].append(property_dict)
 
         return JsonResponse(response, encoder=PintJSONEncoder)
 
@@ -658,7 +537,7 @@ class PropertyViewSet(GenericViewSet):
                 done_searching = False
                 while not done_searching:
                     if (log.parent1_id is None and log.parent2_id is None) or \
-                            log.name == 'Manual Edit':
+                        log.name == 'Manual Edit':
                         done_searching = True
                     else:
                         tree = log.parent1
@@ -670,7 +549,7 @@ class PropertyViewSet(GenericViewSet):
                             record = record_dict(log.parent2)
                             history.append(record)
                         elif log.parent2.name == 'System Match' and log.parent2.parent1.name == 'Import Creation' and \
-                                log.parent2.parent2.name == 'Import Creation':
+                            log.parent2.parent2.name == 'Import Creation':
                             # Handle case where an import file matches within itself, and proceeds to match with
                             # existing records
                             record = record_dict(log.parent2.parent2)
@@ -1518,7 +1397,7 @@ class TaxLotViewSet(GenericViewSet):
                 done_searching = False
                 while not done_searching:
                     if (log.parent1_id is None and log.parent2_id is None) or \
-                            log.name == 'Manual Edit':
+                        log.name == 'Manual Edit':
                         done_searching = True
                     elif log.name == 'Merge current state in migration':
                         record = record_dict(log.parent1)